pyproject.toml
[project]
name = "fast-accounts-local"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "nicegui>=3.6.1",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
    "pywebview>=6.1",
    "supabase>=2.27.3",
]
